refactor(scripts): extract onsite-search PostHog fetcher to shared lib (issue 4301 prep)
profession-keyword-opportunities.mjs's fetchOnsiteSearchTerms() HogQL query
+ auth handling move to scripts/lib/posthog-search-terms.mjs so
mine-search-location-gaps.mjs (profession x city search-demand mining,
issue #4301) reuses the identical query/auth instead of a second copy
(AGENTS.md non-negotiable #6). Verified with --skip-posthog: byte-identical
report output, exit 0.
